grpc-server/requirements.txt
fastnumbers==3.0.0
grpcio==1.31.0
grpcio-tools==1.31.0
protobuf==3.12.4
six==1.15.0
